# Optimize execution of `ZPure`