… core-tests/shared/src/test/scala/zio/prelude/fx/ZPureSpec.scala